Add endpoint methods and tests for finding free cabins and creating a booking

+++ ssbd02/src/main/java/pl/lodz/p/it/ssbd2021/ssbd02/ejb/mop/facades/interfaces/BookingFacadeLocal.java
@@ -50,7 +50,7 @@ public interface BookingFacadeLocal extends AbstractFacadeInterface<Booking> {
      * @param cruise Encja typu {@link Cruise}
      * @return Suma osób zajmujących miejsce na pokładzie promu danego rejsu.
      */
-    int getSumNumberOfPeopleByCruise(Cruise cruise);
+    Long getSumNumberOfPeopleByCruise(Cruise cruise);
 
     /**
      * Metoda sumująca ilosć miejsca zajętą przez pojazdy na promie dla danego rejsu
